backend/requirements.txt
langchain_chroma==1.0.0
langchain_classic==1.0.0
langchain_community==0.4.1
langchain_groq==1.0.0
langchain_huggingface==1.0.1
langchain_text_splitters==1.0.0
python-dotenv==1.2.1

fastapi~=0.121.1
dotenv~=0.9.9

# Auth & security
python-jose[cryptography]==3.5.0
python-multipart>=0.0.20
passlib[bcrypt]==1.7.4
# Pin bcrypt to a version compatible with passlib to avoid '__about__' error
bcrypt==4.0.1
httpx>=0.26.0
chromadb>=0.5.5
sentence-transformers>=2.2.2